+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### 🚀 Features
+- Adding resource for toybox to designate and undesignate toy for return.
+- Adding resource for products get products with query params.
 
 ## [0.0.7] - 2020-04-09
 ### 🚀 Features